Reorganize repository mutations
