+++ app/graphql/mutations/repository/save_repository_mutation.rb
@@ -1,0 +1,34 @@
+# frozen_string_literal: true
+
+module Mutations
+  module Repository
+    SaveRepositoryMutation = GraphQL::Field.define do
+      type Types::RepositoryType
+      description 'Updates a repository'
+
+      argument :id, !types.ID, as: :slug do
+        description 'ID of the repository to update'
+      end
+
+      argument :data, !Types::Repository::ChangesetType do
+        description 'Updated fields of the repository'
+      end
+
+      resource(lambda do |_root, arguments, _context|
+        RepositoryCompound.find(slug: arguments[:slug])
+      end)
+      resolve SaveRepositoryResolver.new
+    end
+
+    # GraphQL mutation to update an repository
+    class SaveRepositoryResolver
+      def call(repository, arguments, _context)
+        params = arguments[:data].to_h.compact
+        params['public_access'] = params.delete('visibility') == 'public'
+        params['description'] = nil if params['description'].empty?
+        repository.update(params)
+        repository
+      end
+    end
+  end
+end
